extract event subscriber tests
