+++ tests/Events/EventsDispatcherTest.php
@@ -373,40 +373,11 @@ class EventsDispatcherTest extends TestCase
         unset($_SERVER['__event.test1']);
         unset($_SERVER['__event.test2']);
     }
-
-    public function testEventSubscribers()
-    {
-        $d = new Dispatcher($container = m::mock(Container::class));
-        $subs = m::mock(ExampleSubscriber::class);
-        $subs->shouldReceive('subscribe')->once()->with($d);
-        $container->shouldReceive('make')->once()->with(ExampleSubscriber::class)->andReturn($subs);
-
-        $d->subscribe(ExampleSubscriber::class);
-        $this->assertTrue(true);
-    }
-
-    public function testEventSubscribeCanAcceptObject()
-    {
-        $d = new Dispatcher();
-        $subs = m::mock(ExampleSubscriber::class);
-        $subs->shouldReceive('subscribe')->once()->with($d);
-
-        $d->subscribe($subs);
-        $this->assertTrue(true);
-    }
 }
 
 class ExampleEvent
 {
     //
-}
-
-class ExampleSubscriber
-{
-    public function subscribe($e)
-    {
-        //
-    }
 }
 
 interface SomeEventInterface
